+++ src/Persistence/Migrations/Full/Full_sqlite.sql
@@ -34,6 +34,8 @@ CREATE TABLE MigrationLog
 -------------------
 -- Account Table --
 -------------------
+-- NOTE: Any changes to the Account schema must be documented in the
+-- "Account Privacy" section of docs/accounts.md.
 
 CREATE TABLE Account
 (
@@ -47,12 +49,16 @@ CREATE UNIQUE INDEX Account_Username_Index ON Account (username);
 ----------------------------
 -- Account Authentication --
 ----------------------------
+-- NOTE: Any changes to the Account_Authentication schema must be documented
+-- in the "Account Privacy" section of docs/accounts.md.
 
 CREATE TABLE Account_Authentication
 (
     id             BLOB PRIMARY KEY NOT NULL,
 	password_salt  BLOB NOT NULL,
 	password_hash  BLOB NOT NULL,
+	opslimit       INTEGER NOT NULL,
+	memlimit       INTEGER NOT NULL,
 	FOREIGN KEY (id) REFERENCES Account(id)
 );
 
@@ -108,9 +114,24 @@ CREATE TABLE Position
 CREATE INDEX Position_Xyz_Index ON Position (x, y, z);
 
 
-----------------------
--- Full Entity View --
-----------------------
+--------------------------------------
+-- Account With Authentication View --
+--------------------------------------
+
+CREATE VIEW AccountWithAuthentication AS
+	SELECT Account.id AS id,
+	       Account.username AS username,
+	       Account_Authentication.salt AS salt,
+		   Account_Authentication.hash AS hash,
+		   Account_Authentication.opslimit AS opslimit,
+		   Account_Authentication.memlimit AS memlimit
+	FROM Account
+	INNER JOIN Account_Authentication ON Account.id = Account_Authentication.id;
+
+
+---------------------------------
+-- Entity With Components View --
+---------------------------------
 
 CREATE VIEW EntityWithComponents AS
     SELECT Entity.id AS id,